v24.9.258 RC1
- added PerformanceStat instead of CpuPerformance class
- CMake: unity build
- Profiler: add GeneralProfiler and RemoteGeneralProfiler with cpu usage, mem usage, etc
- Serializing: Bit packing moved to separate BitSerializer
- Vulkan: add VK_NV_clip_space_w_scaling
- Vulkan: add VK_EXT_subgroup_size_control
- add AE_LICENSE_* to shader code, allow to disable code with unsupported license

+++ AE/engine/src/graphics/Vulkan/VResourceManager.cpp
@@ -167,7 +167,7 @@ namespace AE::Graphics
 
 		VFramebuffer::Key	key{ StructView<ImageViewID>{ desc.attachments.GetValueArray(), &RenderPassDesc::Attachment::imageView },
 								 rp_id,
-								 uint3{ desc.area.Width(), desc.area.Height(), desc.layerCount.Get() }};
+								 uint3{ uint(desc.area.Width()), uint(desc.area.Height()), desc.layerCount.Get() }};
 
 		// find in cache
 		{
